refactoring runmat-config for maintainability
