+++ crates/runmat-config/src/schema/defaults.rs
@@ -1,0 +1,3 @@
+pub(crate) fn default_true() -> bool {
+    true
+}
