+++ crates/runmat-config/src/schema/mod.rs
@@ -1,0 +1,25 @@
+mod accelerate;
+mod defaults;
+mod execution;
+mod kernel;
+mod language;
+mod logging;
+mod plotting;
+mod root;
+mod telemetry;
+
+pub use accelerate::{
+    AccelPowerPreference, AccelerateConfig, AccelerateProviderPreference, AutoOffloadConfig,
+    AutoOffloadLogLevel,
+};
+pub use execution::{GcConfig, GcPreset, JitConfig, JitOptLevel, RuntimeConfig};
+pub use kernel::{KernelConfig, KernelPorts};
+pub use language::{error_namespace_for_language_compat, LanguageCompatMode, LanguageConfig};
+pub use logging::{LogLevel, LoggingConfig};
+pub use plotting::{
+    ExportConfig, ExportFormat, GuiConfig, JupyterConfig, JupyterOutputFormat,
+    JupyterPerformanceConfig, JupyterStaticConfig, JupyterWidgetConfig, PlotBackend, PlotMode,
+    PlottingConfig,
+};
+pub use root::RunMatConfig;
+pub use telemetry::{TelemetryConfig, TelemetryDrainMode};
